+++ src/Utility/Cypress.php
@@ -60,7 +60,7 @@ final class Cypress {
     if (is_null($this->targetElement)) {
       throw new \RuntimeException("You must call ::tag() first.");
     }
-    $test_id = empty($element) ? $this->bemBlock() : $this->bemElement($element);
+    $test_id = $this->bemElement($element);
     if (is_array($this->targetElement)) {
 
       // We assume we have a render array.